[ci skip] ***NO_CI*** Add script to generate ROOT patches for clang
The script uses the current approach of ROOT to manage the patches to
LLVM. ROOT stores a tag file to sync the builtin LLVM with a fork of the
project, which now represents the source of truth for the diffs with
upstream. Patches can be easily generated by comparing the LLVM tag stored
in the ROOT repo with the upstream tag. At this moment, the tags used are:

* LLVM upstream: `llvmorg-13.0.1`
* ROOT: `v6-30-02`

In line with the latest ROOT version for which patches are being
generated. In particular, the latter is used to then checkout the
file with the correct tag of the LLVM fork.